fix: avoid picking the first key's value (indeterministic)
